+++ src/main/java/com/mraof/minestuck/block/MSHorizontalDirectionalBlock.java
@@ -1,0 +1,49 @@
+package com.mraof.minestuck.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+
+import javax.annotation.Nullable;
+
+/**
+ * A copy of {@link HorizontalDirectionalBlock} that is no longer abstract
+ */
+public class MSHorizontalDirectionalBlock extends HorizontalDirectionalBlock
+{
+	
+	protected MSHorizontalDirectionalBlock(Properties properties)
+	{
+		super(properties);
+		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH));
+	}
+	
+	@Override
+	public BlockState rotate(BlockState state, LevelAccessor level, BlockPos pos, Rotation direction)
+	{
+		return state.setValue(FACING, direction.rotate(state.getValue(FACING)));
+	}
+	
+	@Override
+	public BlockState mirror(BlockState state, Mirror mirrorIn)
+	{
+		return state.setValue(FACING, mirrorIn.mirror(state.getValue(FACING)));
+	}
+	
+	@Nullable
+	@Override
+	public BlockState getStateForPlacement(BlockPlaceContext context)
+	{
+		return this.defaultBlockState().setValue(FACING, context.getClickedFace());
+	}
+	
+	@Override
+	protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder)
+	{
+		builder.add(FACING);
+	}
+}