+++ src/main/java/com/mraof/minestuck/data/recipe/MinestuckCombinationsProvider.java
@@ -852,6 +852,8 @@ public final class MinestuckCombinationsProvider
 		CombinationRecipeBuilder.of(MSBlocks.PINK_STONE.get()).input(Tags.Items.STONE).and().namedInput(Items.PINK_DYE).build(consumer);
 		CombinationRecipeBuilder.of(MSBlocks.BROWN_STONE.get()).input(Tags.Items.STONE).and().namedInput(Items.BROWN_DYE).build(consumer);
 		CombinationRecipeBuilder.of(MSBlocks.GREEN_STONE.get()).input(Tags.Items.STONE).and().namedInput(Items.GREEN_DYE).build(consumer);
+		CombinationRecipeBuilder.of(MSBlocks.CARVED_LOG.get()).input(ItemTags.LOGS).or().input(Items.HONEY_BOTTLE).build(consumer); //TODO: replace honey bottle with amber resin
+		CombinationRecipeBuilder.of(MSBlocks.CARVED_WOODEN_LEAF.get()).input(ItemTags.LEAVES).or().input(Items.BOWL).build(consumer);
 		CombinationRecipeBuilder.of(MSBlocks.UNCARVED_WOOD.get()).input(ItemTags.LOGS).and().input(Items.SMOOTH_STONE).build(consumer);
 		CombinationRecipeBuilder.of(MSBlocks.CARVED_HEAVY_PLANKS.get()).input(ItemTags.LOGS).and().input(Items.STONE_BRICKS).build(consumer);
 		CombinationRecipeBuilder.of(MSBlocks.CARVED_KNOTTED_WOOD.get()).input(ItemTags.LOGS).and().input(Items.CHISELED_STONE_BRICKS).build(consumer);