+++ src/main/java/com/mraof/minestuck/util/MSTags.java
@@ -48,6 +48,7 @@ public class MSTags
 		public static final TagKey<Block> ASPECT_LADDERS = tag("ladders/aspect");
 		public static final TagKey<Block> SHADEWOOD_LOGS = tag("logs/shadewood");
 		public static final TagKey<Block> SHADEWOOD_LEAVES = tag("leaves/shadewood");
+		public static final TagKey<Block> WOOD_TERRAIN_BLOCKS = tag("terrain/wood_terrain_blocks");
 		public static final TagKey<Block> CRUXITE_ORES = tag("ores/cruxite");
 		public static final TagKey<Block> URANIUM_ORES = tag("ores/uranium");
 		public static final TagKey<Block> COAL_ORES = tag("ores/coal");